Add FoldersList modal's children component
