+++ src/globalComponents/Modal/ModalChildren.module.css
@@ -1,0 +1,64 @@
+.formContainer {
+  width: 280px;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  margin-top: 10px;
+}
+
+.formInput {
+  width: 100%;
+  padding: 15px 18px;
+  border: 1px solid var(--primary);
+  border-radius: 10px;
+  color: var(--black);
+  font-size: 16px;
+  font-weight: 400;
+}
+
+.formButton {
+  width: 100%;
+  padding: 16px 20px;
+  border-radius: 10px;
+  border: none;
+  background: var(--primary-gradient);
+  font-size: 16px;
+  font-weight: 600;
+  color: var(--white);
+}
+
+.redBackground {
+  background: none;
+  background-color: var(--red);
+}
+
+.shareBox {
+  width: 280px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 32px;
+  margin-top: 10px;
+}
+
+.shareButton {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 10px;
+  font-size: 13px;
+  font-weight: 400;
+  color: var(--black);
+}
+
+.title {
+  font-size: 20px;
+  font-weight: 700;
+  color: var(--black);
+}
+
+.subTitle {
+  font-size: 14px;
+  font-weight: 400;
+  color: var(--gray-600);
+}
