+++ src/pages/FolderPage/components/FoldersList/index.jsx
@@ -2,78 +2,10 @@ import { useState } from "react";
 import styles from "./FoldersList.module.css";
 import { FaPencilAlt, FaRegShareSquare, FaRegTrashAlt } from "react-icons/fa";
 import Modal from "../../../../globalComponents/Modal";
-import kakao from "../../../../assets/kakaotalk.png";
-import facebook from "../../../../assets/facebook.png";
-import share from "../../../../assets/share.png";
-
-function FolderAddForm() {
-  return (
-    <>
-      <h1 className={styles.title}>폴더 추가</h1>
-      <form className={styles.formContainer}>
-        <input
-          className={styles.formInput}
-          type="text"
-          placeholder="내용 입력"
-        />
-        <button className={styles.formButton}>추가하기</button>
-      </form>
-    </>
-  );
-}
-
-function SocialShareBox() {
-  return (
-    <>
-      <h1 className={styles.title}>폴더 공유</h1>
-      <h2 className={styles.subTitle}>폴더명</h2>
-      <div className={styles.shareBox}>
-        <div className={styles.shareButton}>
-          <img src={kakao} alt="kakaoBtn" />
-          <span>카카오톡</span>
-        </div>
-        <div className={styles.shareButton}>
-          <img src={facebook} alt="facebookBtn" />
-          <span>페이스북</span>
-        </div>
-        <div className={styles.shareButton}>
-          <img src={share} alt="shareBtn" />
-          <span>링크 복사</span>
-        </div>
-      </div>
-    </>
-  );
-}
-
-function FolderEditForm() {
-  return (
-    <>
-      <h1 className={styles.title}>폴더 이름 변경</h1>
-      <form className={styles.formContainer}>
-        <input
-          className={styles.formInput}
-          type="text"
-          placeholder="유용한 팁"
-        />
-        <button className={styles.formButton}>변경하기</button>
-      </form>
-    </>
-  );
-}
-
-function FolderDeleteForm() {
-  return (
-    <>
-      <h1 className={styles.title}>폴더 삭제</h1>
-      <h2 className={styles.subTitle}>폴더명</h2>
-      <form className={styles.formContainer}>
-        <button className={`${styles.formButton} ${styles.redBackground}`}>
-          삭제하기
-        </button>
-      </form>
-    </>
-  );
-}
+import FolderAddForm from "./components/FolderAddForm";
+import SocialShareBox from "./components/SocialShareBox";
+import FolderEditForm from "./components/FolderEditForm";
+import FolderDeleteForm from "./components/FolderDeleteForm";
 
 function FoldersList({
   handleClick,
